Moved config funcs to config.go
